Added solutions to Chapter 3.
Some of the primitive functions are not accessible yet due to the
fact that the evaluator does not know how to handle some important
cases yet. For instance, eval doesn't know what to do with Atoms.

There is a convenience function called "repl". It can be used in
ghci to have a little scheme repl going. It is a bit annoying because
it doesn't handle back-space--my most used key.